chore: encapsulate prompt_toolkit-dependant code.
