--- conftool/form/question.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class Question:
-    description: str
-    required: bool
-    prompt: str
-    datatype: type
